tws/requirements.txt
APScheduler==3.11.1
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.13.0
frozendict==2.4.7
idna==3.11
lxml==6.0.2
moomoo_api==10.0.6008
multitasking==0.0.12
numpy==2.3.5
pandas==2.3.3
peewee==4.0.2
platformdirs==4.9.4
protobuf==3.20.3
pycparser==3.0
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
requests==2.32.5
simplejson==3.20.2
six==1.17.0
soupsieve==2.8.3
twstock==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.1
websockets==16.0
yfinance==1.2.0
